fix(filters): enforce content filters on every recommendation surface
Spec §46.4 requires filters on "every query, feed, recommendation, and
notification pipeline" and §46.7.1 restates it as an invariant. They were
enforced in exactly one place: the site search. A reader who blocked a tag
still saw it in /api/v1/discovery, served by three engines.

All three engines now apply the exclusion, inside the statement and before
the limit. Two shapes of work id are involved, so the predicate takes a
work-id *expression* rather than a table name: the site search correlates on
`works.id`, two engines on `w.id`, and the media-reference engine -- which
selects work_media_references -- on `wmr.work_id`. Correlating on the wrong
one is `no such column: works.id` on PostgreSQL.

for_pseud is the only filter lookup. Which pseud a session acts as is
sessions.active_pseud_id, a session property; an account can have several
pseuds, so there is deliberately no account-keyed lookup that guesses one.

Three bugs fixed along the way, all PostgreSQL-only:

- Content filters did nothing on PG. work_tag_values' PG arm used a bare `?`,
  not a bind parameter; the error was swallowed by unwrap_or_default() and
  every reader's filters were silently empty. Same in four settings.rs reads.
- Filtering ran after LIMIT, so a filtered page came back short. A test
  distinguishes the two implementations where the old one could not.
- A `?`/`$n` bind cast to uuid against a TEXT column: six sites, which is
  what four of the five standing PG test failures were. Covered by a new gate.

scripts/check-pg-uuid-casts.py derives the column types from the migrations
and is wired into CI next to the placeholder checker. It resolves table
aliases; without that it reports 81 findings, nearly all false.

Tests: 278 passed / 0 failed on PostgreSQL, up from 272/5. New coverage is
verified to go red by mutation, not just to pass. Clear the four pre-existing
clippy warnings in the tree while here.

+++ crates/db/src/translation.rs
@@ -314,8 +314,12 @@ async fn lookup_memory_postgres(
     Ok(row.map(|r| {
         (
             r.get::<String, _>("target_text"),
-            r.get::<i32, _>("quality_bp") as i64,
-            r.get::<i32, _>("shared") != 0,
+            // The SELECT casts both to BIGINT, so the reader is `i64`. Reading
+            // them as `i32` is `INT4 is not compatible with INT8` -- a decode
+            // error, not a truncation. The SQLite arm selects the columns
+            // uncast and correctly reads `i32`.
+            r.get::<i64, _>("quality_bp"),
+            r.get::<i64, _>("shared") != 0,
         )
     }))
 }
